[WIP] Wrap Remark, Json nodes in ContentPage
so that we have a uniform data structure to relate to in the UI no matter whether it is a new or old, WordPress post/page.

Since Remark Transformer creates important fields as graphql-query-time custom (lazy) plugin fields, we hack around and
reuse the resolvers it defines.

See TODOs and FIXMEs (include remark plugins, add other fields, add Json fields)